requirements.txt
django==1.8
python-social-auth
djangorestframework
Pillow
django-rest-auth[with_social]
django-rest-swagger
